Rename Boss for CSPWorkerPool, mv it to block module
CSP stands for communicating sequencial processes which is the model
that is being encapsulated. Move the implementation to the block module
as the client does not get all the results back until the very end. This
will allow us to provide a streaming model, where the client gets the
results back immediately as they are being obtained.

Signed-off-by: Jordi Carrillo <jordilin@gmail.com>