Add CSPStreamWorkerPool
The CSPStreamWorkerPool streams the results from the callback function
back to the client. It shares similar functionality as the CSPWorkerPool
except that the internal Worker holds a sender channel to send the
results back to the client.

Add an example to demonstrate how the sending and receival of messages
are interweaved.

Signed-off-by: Jordi Carrillo <jordilin@gmail.com>